fix: ci, decorate name converter
